Make the verification test pass.
The current implementation of "exponential modulo" is
not working properly; I replaced it with another one
I found.

Test: $ python2 dsa.py

I found this while looking for DSA implementations
for the Matasano crypto challenges. I also made it
compatible with python3 (but not with both python2
and python3), let me know if you are interested.